Create an assembler
Nand2Tetris Project 06

Why:
We would like to have an working assembler for the hack computer,
assembling hack assembly into hack machine code.

This commit:
Creates a hack assembler written in elixir. It can handle both the
translations of a-instructions and c-instruction into 16bit machine
commands and the higher level niceties such as symbol and label
resolution.

See the mix.exs file for info on how to run the program.