Create a VM to hack translator
Nand2Tetris P07

Why:
We would like to be able to translate from intermediate language for the
hack VM to hack machine language.

This PR:
Creates such a translator in elixir. It can handle arithmetic commands
and memory access commands for the hack vm, but not branching logic or
function declarations.

See mix.exs for instructions on how to run the program.